vectorize movegen
https://87flowers.com/byteboard-attack-tables-1/

--------------------------------------------------
Results of dev vs base (1+0.01, 1t, 64MB, UHO_Lichess_4852_v1.epd):
Elo: 8.27 +/- 4.05, nElo: 14.83 +/- 7.27
LOS: 100.00 %, DrawRatio: 47.40 %, PairsRatio: 1.20
Games: 8784, Wins: 2439, Losses: 2230, Draws: 4115, Points: 4496.5 (51.19 %)
Ptnml(0-2): [71, 980, 2082, 1187, 72], WL/DD Ratio: 1.14
LLR: 2.90 (100.3%) (-2.25, 2.89) [0.00, 3.00]
--------------------------------------------------

+++ lib/search/attention.rs
@@ -33,7 +33,7 @@ impl Stat for Nodes {
 }
 
 /// Measures the effort spent searching a root [`Move`].
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Zeroable)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Zeroable)]
 #[debug("Attention")]
 pub struct Attention(Butterfly<Nodes>);
 
@@ -48,6 +48,6 @@ impl Attention {
     #[inline(always)]
     #[cfg_attr(feature = "no_panic", no_panic::no_panic)]
     pub fn nodes(&mut self, m: Move) -> &mut Nodes {
-        &mut self.0[m.whence() as usize][m.whither() as usize]
+        &mut self.0[m.whence()][m.whither()]
     }
 }
